pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "servotest"
version = "0.1.9"
description = "A simple command line tool for testing and moving servos"
authors = [
    {name="Neo Yang", email="neoxyyang11@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Public Domain",
    "Operating System :: OS Independent",
    "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
    "pyserial>=3.5",
    "colorama>=0.4.6",
]

[project.urls]
"Homepage" = "https://github.com/N30Yang/Servotest"
"Bug Tracker" = "https://github.com/N30Yang/Servotest/issues"

[project.scripts]
servotest = "servotest.main:start_app"

[tool.hatch.build.targets.wheel]
packages = ["src/servotest"]